Data.Ini

Description

Clean configuration files in the INI format.

Format rules and recommendations:

The : syntax is space-sensitive.

Keys are case-sensitive.

Lower-case is recommended.

Values can be empty.

Keys cannot key separators, section delimiters, or comment markers.

Comments must start at the beginning of the line and start with ; or #.

An example configuration file:

# Some comment.
[SERVER]
port=6667
hostname=localhost
; another comment here
[AUTH]
user: hello
pass: world
salt:

Parsing example:

>>> parseIni "[SERVER]\nport: 6667\nhostname: localhost"
Right (Ini {unIni = fromList [("SERVER",fromList [("hostname","localhost"),("port","6667")])]})

Reading

readIniFile :: FilePath -> IO (Either String Ini)

Parse an INI file.

parseIni :: Text -> Either String Ini

Parse an INI config.

lookupValue

Arguments

:: Text
Section name

-> Text
Key

-> Ini

-> Either String Text

Lookup one value in the config.

Example:

>>> parseIni "[SERVER]\nport: 6667\nhostname: localhost" >>= lookupValue "SERVER" "hostname"
Right "localhost"

lookupArray

Arguments

:: Text
Section name

-> Text
Key

-> Ini

-> Either String [Text]

Lookup one value in the config.

Example:

>>> parseIni "[SERVER]\nport: 6667\nhostname: localhost" >>= lookupValue "SERVER" "hostname"
Right "localhost"

readValue

Arguments

:: Text
Section name

-> Text
Key

-> (Text -> Either String (a, Text))

-> Ini

-> Either String a

Read a value using a reader from Data.Text.Read.

readArray

Arguments

:: Text
Section name

-> Text
Key

-> (Text -> Either String (a, Text))

-> Ini

-> Either String [a]

Read an array of values using a reader from Data.Text.Read.

parseValue

Arguments

:: Text
Section name

-> Text
Key

-> Parser a

-> Ini

-> Either String a

Parse a value using a reader from Data.Attoparsec.Text.

sections :: Ini -> [Text]

Get the sections in the config.

Example:

>>> sections <$> parseIni "[SERVER]\nport: 6667\nhostname: localhost"
Right ["SERVER"]

keys

Arguments

:: Text
Section name

-> Ini

-> Either String [Text]

Get the keys in a section.

Example:

>>> parseIni "[SERVER]\nport: 6667\nhostname: localhost" >>= keys "SERVER"
Right ["hostname","port"]

Writing

printIni :: Ini -> Text

Print an INI config.

writeIniFile :: FilePath -> Ini -> IO ()

Print the INI config to a file.

Advanced writing

data KeySeparator

Either : or =.

Constructors

ColonKeySeparator

EqualsKeySeparator

data WriteIniSettings

Settings determining how an INI file is written.

Constructors

WriteIniSettings

Fields

writeIniKeySeparator :: KeySeparator

defaultWriteIniSettings :: WriteIniSettings

The default settings for writing INI files.

printIniWith :: WriteIniSettings -> Ini -> Text

Print an INI config.

writeIniFileWith :: WriteIniSettings -> FilePath -> Ini -> IO ()

Print the INI config to a file.

Types

data Ini

An INI configuration.

Constructors

Ini

Fields

iniSections :: HashMap Text [(Text, Text)]

iniGlobals :: [(Text, Text)]

unIni :: Ini -> HashMap Text (HashMap Text Text)

Parsers

iniParser :: Parser Ini

Parser for an INI.

sectionParser :: Parser (Text, [(Text, Text)])

A section. Format: [foo]. Conventionally, [FOO].

keyValueParser :: Parser (Text, Text)

A key-value pair. Either foo: bar or foo=bar.